+++ proxy-template.php
@@ -9,11 +9,12 @@ use Brick\Geo\Exception\CoordinateSystemException;
 use Brick\Geo\Exception\InvalidGeometryException;
 use Brick\Geo\Exception\UnexpectedGeometryException;
 use Brick\Geo\Geometry;
+use Brick\Geo\_CLASSNAME_;
 
 /**
  * Proxy class for _CLASSNAME_.
  */
-class _CLASSNAME_Proxy extends _FQCN_ implements ProxyInterface
+class _CLASSNAME_Proxy extends _CLASSNAME_ implements ProxyInterface
 {
     /**
      * The WKT or WKB data.
@@ -33,7 +34,7 @@ class _CLASSNAME_Proxy extends _FQCN_ implements ProxyInterface
     /**
      * The underlying geometry, or NULL if not yet loaded.
      */
-    private ?_FQCN_ $proxyGeometry = null;
+    private ?_CLASSNAME_ $proxyGeometry = null;
 
     /**
      * @param string $data     The WKT or WKB data.
@@ -58,8 +59,8 @@ class _CLASSNAME_Proxy extends _FQCN_ implements ProxyInterface
     private function load() : void
     {
         $this->proxyGeometry = $this->isProxyBinary
-            ? _FQCN_::fromBinary($this->proxyData, $this->proxySRID)
-            : _FQCN_::fromText($this->proxyData, $this->proxySRID);
+            ? _CLASSNAME_::fromBinary($this->proxyData, $this->proxySRID)
+            : _CLASSNAME_::fromText($this->proxyData, $this->proxySRID);
     }
 
     public function isLoaded() : bool