+++ src/Proxy/MultiPolygonProxy.php
@@ -9,11 +9,12 @@ use Brick\Geo\Exception\CoordinateSystemException;
 use Brick\Geo\Exception\InvalidGeometryException;
 use Brick\Geo\Exception\UnexpectedGeometryException;
 use Brick\Geo\Geometry;
+use Brick\Geo\MultiPolygon;
 
 /**
  * Proxy class for MultiPolygon.
  */
-class MultiPolygonProxy extends \Brick\Geo\MultiPolygon implements ProxyInterface
+class MultiPolygonProxy extends MultiPolygon implements ProxyInterface
 {
     /**
      * The WKT or WKB data.
@@ -33,7 +34,7 @@ class MultiPolygonProxy extends \Brick\Geo\MultiPolygon implements ProxyInterfac
     /**
      * The underlying geometry, or NULL if not yet loaded.
      */
-    private ?\Brick\Geo\MultiPolygon $proxyGeometry = null;
+    private ?MultiPolygon $proxyGeometry = null;
 
     /**
      * @param string $data     The WKT or WKB data.
@@ -58,8 +59,8 @@ class MultiPolygonProxy extends \Brick\Geo\MultiPolygon implements ProxyInterfac
     private function load() : void
     {
         $this->proxyGeometry = $this->isProxyBinary
-            ? \Brick\Geo\MultiPolygon::fromBinary($this->proxyData, $this->proxySRID)
-            : \Brick\Geo\MultiPolygon::fromText($this->proxyData, $this->proxySRID);
+            ? MultiPolygon::fromBinary($this->proxyData, $this->proxySRID)
+            : MultiPolygon::fromText($this->proxyData, $this->proxySRID);
     }
 
     public function isLoaded() : bool