+++ src/Proxy/TriangleProxy.php
@@ -9,11 +9,12 @@ use Brick\Geo\Exception\CoordinateSystemException;
 use Brick\Geo\Exception\InvalidGeometryException;
 use Brick\Geo\Exception\UnexpectedGeometryException;
 use Brick\Geo\Geometry;
+use Brick\Geo\Triangle;
 
 /**
  * Proxy class for Triangle.
  */
-class TriangleProxy extends \Brick\Geo\Triangle implements ProxyInterface
+class TriangleProxy extends Triangle implements ProxyInterface
 {
     /**
      * The WKT or WKB data.
@@ -33,7 +34,7 @@ class TriangleProxy extends \Brick\Geo\Triangle implements ProxyInterface
     /**
      * The underlying geometry, or NULL if not yet loaded.
      */
-    private ?\Brick\Geo\Triangle $proxyGeometry = null;
+    private ?Triangle $proxyGeometry = null;
 
     /**
      * @param string $data     The WKT or WKB data.
@@ -58,8 +59,8 @@ class TriangleProxy extends \Brick\Geo\Triangle implements ProxyInterface
     private function load() : void
     {
         $this->proxyGeometry = $this->isProxyBinary
-            ? \Brick\Geo\Triangle::fromBinary($this->proxyData, $this->proxySRID)
-            : \Brick\Geo\Triangle::fromText($this->proxyData, $this->proxySRID);
+            ? Triangle::fromBinary($this->proxyData, $this->proxySRID)
+            : Triangle::fromText($this->proxyData, $this->proxySRID);
     }
 
     public function isLoaded() : bool