+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ✨ **Improvements**
 
 - Removed outdated suggestion in `composer.json`
+- Minor proxy code style improvements
 
 ## [0.8.0](https://github.com/brick/geo/releases/tag/0.8.0) - 2022-10-11